refactor: moved filesys code to own module
