+++ src/core/mod.rs
@@ -1,0 +1,3 @@
+pub mod archives;
+pub mod filesys;
+pub mod magic;
